Cleanup redundant API integration files
